Delete rustdoc::doctree
